+++ docs/html/search/classes_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['kstring_60',['KString',['../classace__common_1_1KString.html',1,'ace_common']]]
+  ['kstring_63',['KString',['../classace__common_1_1KString.html',1,'ace_common']]]
 ];