+++ docs/html/search/files_5.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['url_5fencoding_2eh_72',['url_encoding.h',['../url__encoding_8h.html',1,'']]]
+  ['printfto_2eh_73',['printfTo.h',['../printfTo_8h.html',1,'']]],
+  ['printpadto_2eh_74',['printPadTo.h',['../printPadTo_8h.html',1,'']]],
+  ['pstrings_2eh_75',['pstrings.h',['../pstrings_8h.html',1,'']]]
 ];