+++ docs/html/search/variables_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['buf_5f_113',['buf_',['../classace__common_1_1PrintStrBase.html#a0b8ea388f529134fdfc0cde78d866b9a',1,'ace_common::PrintStrBase']]]
+  ['buf_5f_119',['buf_',['../classace__common_1_1PrintStrBase.html#a0b8ea388f529134fdfc0cde78d866b9a',1,'ace_common::PrintStrBase']]]
 ];